remove /src directory
